# DRY conditions.go / small perf boost