--- a/libbeat/autodiscover/config.go
+++ b/libbeat/autodiscover/config.go
@@ -19,7 +19,7 @@ package autodiscover
 
 import (
 	"github.com/elastic/beats/libbeat/common"
-	"github.com/elastic/beats/libbeat/processors"
+	"github.com/elastic/beats/libbeat/conditions"
 )
 
 // Config settings for Autodiscover
@@ -39,6 +39,6 @@ type BuilderConfig struct {
 
 // AppenderConfig settings
 type AppenderConfig struct {
-	Type            string                      `config:"type"`
-	ConditionConfig *processors.ConditionConfig `config:"condition"`
+	Type            string             `config:"type"`
+	ConditionConfig *conditions.Config `config:"condition"`
 }