--- a/libbeat/conditions/and.go
+++ b/libbeat/conditions/and.go
@@ -15,13 +15,30 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package processors
+package conditions
 
-import (
-	"github.com/elastic/beats/libbeat/common"
-)
+// And is a compound condition that combines multiple conditions with logical AND.
+type And []Condition
 
-type PluginConfig []map[string]*common.Config
+// NewAndCondition builds this condition from a slice of Condition objects.
+func NewAndCondition(conditions []Condition) And {
+	return And(conditions)
+}
 
-// fields that should be always exported
-var MandatoryExportedFields = []string{"type"}
+// Check determines whether the given event matches this condition
+func (c And) Check(event ValuesMap) bool {
+	for _, cond := range c {
+		if !cond.Check(event) {
+			return false
+		}
+	}
+	return true
+}
+
+func (c And) String() (s string) {
+	for _, cond := range c {
+		s = s + cond.String() + " and "
+	}
+	s = s[:len(s)-len(" and ")] //delete the last and
+	return s
+}