conditions.go was checking the configured conditions with conditional branching both at
config load time and once per event. This patch removes the redundant per-event checks.

There may be a small perf boost with this patch as well.