@urso 2018-07-02
I'm pretty sure this change will boost performance a lot.

A more idiomatic solution would use interfaces instead of a function pointer. E.g. change `Condition` type to `type Condition interface` and create one implementation per condition type.

Benchmarks would be nice to have for comparing future changes as well.