@andrewvc 2018-07-02
@ruflin @urso thanks for the feedback, glad to make both refactors. I like your proposal of one concrete implementation per condition type. The current logic, based on which fields happen to be set, is harder to understand.