@andrewvc 2018-07-02
This is an in-progress commit, I still have to run benchmarks against it and maybe add more tests. A few small checks may also be failing. It would be great to have feedback on the general direction here at this point however.

This PR is now *slightly* larger , but is, IMHO, much better, having taken @urso 's suggestion of using an interface rather than a function pointer. 

There was a lot of logic in the `conditionals.go` file which is now better broken up, more readable, and function pointer free. Porting these functions to interfaces and leaving them in a monolithic file felt like a waste of effort, since it'd still be hard to read. 

This is a pretty big refactor, I tried to limit it to mostly moving code around around, and kept actual refactors to a bare minimum, so while the number of lines is large, the amount of new logic is minimal. I did have to add a lot of comments to old functions to make Hound happy.

A key change here is the creation of a new `conditionals` package, which encapsulates the conditionals logic. Keeping it all in `processors` seemed like not a great idea. There's more than enough code in conditionals to justify a new package IMHO.

Another significant change was moving the conditional objects to be values rather than pointer receivers to following the style recommendations on the [golang code review comments](https://github.com/golang/go/wiki/CodeReviewComments#receiver-type). `Conditional` concrete types are all either immutable maps , tiny structs, or slices. I'm a golang newbie, so maybe I'm misunderstanding that advice, but it seems like anything large is hidden away in a map or a slice, and the remaining copied values are small.

Looking over the code, we don't have great test coverage here at the unit level, it might be nice to add some more tests, but I'm fine leaving the test coverage as is, which is to say, largely at the functional level. Thoughts @urso && @ruflin ? I think it would be fair to say that's out of scope for this PR at the very least, this is already a quite large one.