@andrewvc 2018-07-03
OK, added some benchmarks, tested between master / this branch. It's a nice little boost for this part of the code, but at the end of the day, I doubt users will perceive a perf improvement here unless it turns out that the efficiency here improves the global story for CPU caching, which is possible, but probably not worth spending time to benchmark (I also doubt it will amount to anything anyone will care about).

With the benchmark data we do have, in the best case, you save 10ns per condition, which is about 10ms per million events with one condition. If you have a lot of those conditions that *could* add up. However, that savings is relatively smaller as the statements grow in complexity.

Raw benchmark data below. This is comparing the benchmarks in the old branch (which I added in https://github.com/andrewvc/beats/blob/dry-cond-old-bench/libbeat/processors/condition_test.go) with the new branch in ( https://github.com/andrewvc/beats/blob/dry-cond/libbeat/processors/conditions/conditions_test.go ).

The simple condition test tests a `has_fields` condition, which is about as simple as it gets. The `CombinedCondition` test tests more complexity with a compound statement.

```
benchmark                        old ns/op     new ns/op     delta
BenchmarkSimpleCondition-8       34.9          21.5          -38.40%
BenchmarkCombinedCondition-8     149           127           -14.77%
```

We should still do this, since the code is much better organized and easy to work with in this PR, and the perf is clearly better.