@ruflin 2018-07-04
For me the biggest win of this change is the actual refactoring with the interfaces as it should make adding future conditions easier.