@urso 2018-07-04
Nice code cleanups. I love finally having an extra package. Conditions are not only used by processors. You might consider to move the package one level up.

The tests still need some cleanup. Current test code makes it quite hard to localize and debug an actual failing test.  E.g. have tests in file per condtion (e.g. `and.go` with tests in `and_test.go`). Plus use `t.Run` for sub tests. So it's easier to find and run the exact test that failed.

Common pattern for table driven tests:

```
func TestX(t *testing.T) {
  type testCase struct {
    // test input, expected output, ...
  }

  cases := map[string]testCase{ // using map here ensures tests are run in random order
    "must succeed": testCase{ ... },
    "must fail": testCase{...}, 
  }

  for name, test := range cases {
    // common global setup/teardown

    test := test // capture loop variable in case tests are run in parallel (via t.Parallel())
    t.Run(name, func(t *testing.T) {
      // test driver code
    })
  }
}
```

Using this pattern, the test name will appear in logs + one can run tests selectively via `go test -run 'X/fail/'`. The `-run` arguments accepts a regular expressions with `/` delimiting test levels.