@urso 2018-07-04
I'm not too worried about the benchmarks here (users can end up with tons of conditionals). The refactoring makes the code much more accessible and maintainable.