@andrewvc 2018-07-05
@urso great point WRT to the tests, I'll split them out later today. Thanks for the advice on test structure as well.