@andrewvc 2018-07-05
I've made the following changes:

1. Refactored the various tests out to separate files + broke up the big combination tests into individual ones. Ultimately, it made more sense to do a function per-test than to go the anonymous route.
2. Renamed conditions.Execute to conditions.Check

I think that we should do some of the additional work in follow up PRs. This means:

1. Refactoring `conditions.NewCondition` to remove the big switch and validate in a more robust way.
2. Potentially making the tests more comprehensive. As I've split out the tests I've noticed that some conditions are tested better than others. I'd like to hear what @urso and @ruflin think about the current state of the tests, but they are at the moment spot checks, and not comprehensive. For this core functionality I think it makes sense to test a range of different conditions instead of the current spot checks.

I would prefer to do this in separate PRs because this one is unwieldy and I'd like to constrain its scope to reorganization. While there is some refactoring in here, it is in order to facilitate reorganization, not to change any actual logic besides the core goal of removing the once-per-event branching on condition configuration.