@andrewvc 2018-07-05
I also just moved the package up a level to be a sibling of processors as @urso suggested.