@ruflin 2018-07-06
@andrewvc There seems to be something off with the formatting. `make fmt` should solve it.